// openmetadata-ui/src/main/resources/ui/src/pages/CustomizablePage/CustomizablePage.interface.ts
/*
 *  Copyright 2023 Collate.
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
 *  You may obtain a copy of the License at
 *  http://www.apache.org/licenses/LICENSE-2.0
 *  Unless required by applicable law or agreed to in writing, software
 *  distributed under the License is distributed on an "AS IS" BASIS,
 *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 */

import { ReactNode } from 'react';
import { Layout } from 'react-grid-layout';

export interface WidgetConfig
  extends Pick<Layout, 'i' | 'x' | 'y' | 'h' | 'w' | 'static' | 'isDraggable'> {
  children?: WidgetConfig[];
  config?: {
    [key: string]: any;
  };
  data?: {
    page: {
      layout: Array<WidgetConfig>;
    };
  };
}

export interface WidgetCommonProps {
  selectedGridSize?: number;
  isEditView?: boolean;
  widgetKey: string;
  dragHandle?: ReactNode;
  handleRemoveWidget?: (widgetKey: string) => void;
  handleLayoutUpdate?: (layout: Layout[]) => void;
  handleSaveLayout?: (layout: WidgetConfig[]) => Promise<void>;
  currentLayout?: Array<WidgetConfig>;
}